MultiUpdate bug fix & Views on their own headers
