+++ Editor/MaterialView.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "PropInspector.h"
+
+namespace ToolKit
+{
+  namespace Editor
+  {
+
+
+  }
+}
