+++ Editor/PropInspector.h
@@ -10,21 +10,22 @@ namespace ToolKit
 {
   namespace Editor
   {
-    void DropZone(uint fallbackIcon,
-                  const String& file,
-                  std::function<void(const DirectoryEntry& entry)> dropAction,
-                  const String& dropName = "");
-
-    void DropSubZone(
-        const String& title,
-        uint fallbackIcon,
-        const String& file,
-        std::function<void(const DirectoryEntry& entry)> dropAction,
-        bool isEditable);
 
     class View
     {
      public:
+      static void DropZone(uint fallbackIcon,
+                    const String& file,
+                    std::function<void(const DirectoryEntry& entry)> dropAction,
+                    const String& dropName = "");
+      static void DropSubZone(
+          const String& title,
+          uint fallbackIcon,
+          const String& file,
+          std::function<void(const DirectoryEntry& entry)> dropAction,
+          bool isEditable);
+      static bool IsTextInputFinalized();
+
       View(const StringView viewName);
       virtual ~View()
       {
@@ -36,49 +37,6 @@ namespace ToolKit
       int m_viewID         = 0;
       TexturePtr m_viewIcn = nullptr;
       const StringView m_viewName;
-    };
-
-    class PrefabView : public View
-    {
-     public:
-      PrefabView();
-      virtual ~PrefabView();
-      virtual void Show();
-
-     private:
-      bool DrawHeader(Entity* ntt, ImGuiTreeNodeFlags flags);
-      void ShowNode(Entity* e);
-
-     private:
-      Entity* m_activeChildEntity = nullptr;
-    };
-
-    class EntityView : public View
-    {
-     public:
-      EntityView();
-      virtual ~EntityView();
-      virtual void Show();
-      virtual void ShowParameterBlock();
-
-     protected:
-      void ShowAnchorSettings();
-    };
-
-    class ComponentView : public View
-    {
-     public:
-      ComponentView();
-      virtual ~ComponentView();
-      virtual void Show();
-    };
-
-    class CustomDataView : public View
-    {
-     public:
-      CustomDataView();
-      virtual ~CustomDataView();
-      virtual void Show();
     };
 
     typedef View* ViewRawPtr;
